# cmd/server/docs/swagger.yaml
basePath: /api/v1
definitions:
  config.ModelEntry:
    properties:
      api_key:
        type: string
      base_url:
        type: string
      chat_model:
        type: string
      embedding_model:
        type: string
      name:
        type: string
      provider:
        type: string
    type: object
...
  handler.modelItem:
    properties:
      active:
        type: boolean
      is_custom:
        type: boolean
      is_local:
        type: boolean
      name:
        type: string
      provider:
        type: string
    type: object
  model.APIEnvelope:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  model.DeleteDocumentResponse:
    properties:
      document_id:
        type: string
      status:
        type: string
    type: object
  model.DocumentItem:
    properties:
      chunk_count:
        type: integer
      content:
        type: string
      created_at:
        type: string
      document_id:
        type: string
    type: object
...
  model.ListDocumentsResponse:
    properties:
      documents:
        items:
          $ref: '#/definitions/model.DocumentItem'
        type: array
      total:
        type: integer
    type: object
...
  model.UploadDocumentResponse:
    properties:
      chunk_count:
        type: integer
      document_ids:
        items:
          type: string
        type: array
      status:
        type: string
    type: object
  service.SkillEntry:
    properties:
      enabled:
        type: boolean
      name:
        type: string
      prompt:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    name: MrLuoooooo
    url: https://github.com/MrLuoooooo/MrLuoooooagent
  description: AI Agent 后端系统。让大模型安全操控本地开发环境——读写文件、执行命令、联网搜索、查询股票、检索知识库。
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://github.com/MrLuoooooo/MrLuoooooagent
  title: GoAgent Pro API
  version: 4.1.2
paths:
...
  /approvals/{approval_id}/decide:
    post:
      consumes:
      - application/json
      description: 接受或拒绝一个待审批的定时任务操作。
      parameters:
      - description: 审批项 ID
        in: path
        name: approval_id
        required: true
        type: string
      - description: 决策（accept=true 接受，false 拒绝）
        in: body
        name: request
        required: true
        schema:
          properties:
            accept:
              type: boolean
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.APIEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.APIEnvelope'
      summary: 审批决策
      tags:
      - 审批
...
  /documents:
    get:
      description: 获取所有已上传的文档列表。
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.APIEnvelope'
            - properties:
                data:
                  $ref: '#/definitions/model.ListDocumentsResponse'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.APIEnvelope'
      summary: 列出文档
      tags:
      - 文档
    post:
      consumes:
      - multipart/form-data
      description: 上传文档文件（.pdf/.txt/.md），自动进行分块、向量化并存入知识库。最大 10MB。
      parameters:
      - description: 上传文件（支持 .pdf / .txt / .md）
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.APIEnvelope'
            - properties:
                data:
                  $ref: '#/definitions/model.UploadDocumentResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.APIEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.APIEnvelope'
      summary: 上传文档
      tags:
      - 文档
  /documents/{document_id}:
    delete:
      description: 按文档 ID 删除已上传的文档及其向量索引。
      parameters:
      - description: 文档 ID
        in: path
        name: document_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.APIEnvelope'
            - properties:
                data:
                  $ref: '#/definitions/model.DeleteDocumentResponse'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.APIEnvelope'
      summary: 删除文档
      tags:
      - 文档
  /models:
    get:
      description: 获取所有可用模型列表。包含配置中的预置模型和用户自定义模型，标记当前活跃模型。
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.APIEnvelope'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/handler.modelItem'
                  type: array
              type: object
      summary: 列出可用模型
      tags:
      - 模型
    post:
      consumes:
      - application/json
      description: 添加一个自定义 LLM 模型到可用模型列表。
      parameters:
      - description: 自定义模型配置
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/config.ModelEntry'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.APIEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.APIEnvelope'
      summary: 添加自定义模型
      tags:
      - 模型
  /models/{name}:
    delete:
      description: 从可用模型列表中删除一个自定义模型。
      parameters:
      - description: 模型名称
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.APIEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.APIEnvelope'
      summary: 删除自定义模型
      tags:
      - 模型
  /models/switch:
    post:
      consumes:
      - application/json
      description: 运行时切换 LLM 模型，无需重启服务。
      parameters:
      - description: 模型名称
        in: body
        name: request
        required: true
        schema:
          properties:
            model:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.APIEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.APIEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.APIEnvelope'
      summary: 切换模型
      tags:
      - 模型
  /skills:
    get:
      description: 获取所有用户自定义技能列表。技能会在 Agent 运行时注入到系统提示词中。
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.APIEnvelope'
      summary: 列出技能
      tags:
      - 技能
    post:
      consumes:
      - application/json
      description: 添加或更新一个用户自定义技能。技能会在 Agent 执行时自动注入提示词。
      parameters:
      - description: 技能配置（name + prompt）
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.SkillEntry'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.APIEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.APIEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.APIEnvelope'
      summary: 添加/更新技能
      tags:
      - 技能
  /skills/{name}:
    delete:
      description: 按名称删除一个用户自定义技能。
      parameters:
      - description: 技能名称
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.APIEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.APIEnvelope'
      summary: 删除技能
      tags:
      - 技能
...
swagger: "2.0"